requirements.txt
discord.py==2.4.0
APScheduler==3.10.4
python-dotenv==1.0.1
PyYAML==6.0.2
pytz==2024.1
